--- src/forms.ts
@@ -1,7 +1,0 @@
-/**
- * Created by sam on 06.03.2017.
- */
-
-
-export const FORM_COPYNUMBER_SUBTYPE_ID = 'copynumber_subtype';
-export const FORM_EXPRESSION_SUBTYPE_ID = 'expression_subtype';
