+++ src/index.ts
@@ -1,11 +1,5 @@
 export * from './menu';
 export * from './views';
-
 export * from './common';
-export * from './constants';
-export * from './extensions';
-export * from './forms';
-export * from './GeneIDTypeDetector';
+export * from './provider';
 export {default} from './phovea';
-export * from './SearchProvider';
-export * from './utils';
