+++ src/provider/forms.ts
@@ -1,0 +1,8 @@
+/**
+ * Created by sam on 06.03.2017.
+ */
+
+export module FormSubtype {
+  export const FORM_COPYNUMBER_SUBTYPE_ID = 'copynumber_subtype';
+  export const FORM_EXPRESSION_SUBTYPE_ID = 'expression_subtype';
+}
