+++ src/scss/_legend.scss
