+++ src/scss/_species.scss
@@ -3,7 +3,7 @@
     content: '';
     display: block;
     height: 0.85em;
-    background: no-repeat center/50px url('./styles/mouse-icon.png');
+    background: no-repeat center/50px url('./mouse-icon.png');
     opacity: 0.1;
     transition: opacity 0.2s ease-in-out;
   }
