+++ src/scss/mouse-icon.png
