+++ src/scss/style.scss
@@ -1,0 +1,6 @@
+// styles for views
+@import "onco_print";
+@import "expression_vs_copynumber";
+@import "species";
+@import "startMenu";
+@import "legend";
