--- src/style.scss
@@ -1,6 +1,0 @@
-// styles for views
-@import "styles/onco_print";
-@import "styles/expression_vs_copynumber";
-@import "styles/species";
-@import "styles/startMenu";
-@import "styles/legend";
