+++ src/views/GeneProxyView.ts
@@ -3,7 +3,7 @@
  */
 
 import {ProxyView} from 'tdp_core/src/views/ProxyView';
-import {createOptions, getSelectedSpecies} from '../common';
+import {SpeciesUtils} from '../common/common';
 import {IFormSelectOption} from 'tdp_core/src/form';
 
 /**
@@ -12,12 +12,12 @@ import {IFormSelectOption} from 'tdp_core/src/form';
 export class GeneProxyView extends ProxyView {
 
   protected getSelectionSelectData(ensgs: string[]): Promise<IFormSelectOption[]> {
-    return createOptions(ensgs, this.selection, this.idType);
+    return SpeciesUtils.createOptions(ensgs, this.selection, this.idType);
   }
 
   protected updateProxyView() {
     const extra: any = this.options.extra;
-    extra.species = getSelectedSpecies();
+    extra.species = SpeciesUtils.getSelectedSpecies();
     super.updateProxyView();
   }
 }
