+++ src/views/index.ts
@@ -4,4 +4,4 @@ export * from './AExpressionVsCopyNumber';
 export * from './AOncoPrint';
 export * from './GeneProxyView';
 export * from './UniProtProxyView';
-export * from './utils';
+export * from './ViewUtils';
